refactor: saveToLocalStorage 関数をlocalStorageContextに移動 不要なファイルを削除
